Move blas.h/cpp and lapack.h/cpp to subdirectory "math" and add missing namespace.
