+++ python/boomer/boosting/_lapack.pxd
@@ -11,7 +11,7 @@ from scipy.linalg.cython_lapack cimport dsysv
 from libcpp.memory cimport unique_ptr, make_unique
 
 
-cdef extern from "cpp/lapack.h" nogil:
+cdef extern from "cpp/math/lapack.h" namespace "boosting" nogil:
 
     ctypedef void (*dsysv_t)(char* uplo, int* n, int* nrhs, double* a, int* lda, int* ipiv, double* b, int* ldb, double* work, int* lwork, int* info)
