+++ python/setup.py
@@ -68,9 +68,9 @@ sources = [
     'boomer/boosting/cpp/losses/loss_label_wise_logistic.cpp',
     'boomer/boosting/cpp/losses/loss_label_wise_squared_error.cpp',
     'boomer/boosting/cpp/losses/loss_example_wise_logistic.cpp',
+    'boomer/boosting/cpp/math/blas.cpp',
+    'boomer/boosting/cpp/math/lapack.cpp',
     'boomer/boosting/cpp/post_processing/shrinkage_constant.cpp',
-    'boomer/boosting/cpp/blas.cpp',
-    'boomer/boosting/cpp/lapack.cpp',
     'boomer/boosting/cpp/rule_evaluation_label_wise.cpp',
     'boomer/boosting/cpp/rule_evaluation_example_wise.cpp',
     'boomer/boosting/cpp/statistics/statistics_label_wise.cpp',
